dev-7567 add image and banner to homepage main content
