+++ src/_scss/pages/homepage/covid-vaccine-link.scss
@@ -1,0 +1,37 @@
+.covid-19-vaccine-link__container {
+  background-color: #0051C4;
+  color: white;
+  padding-left: $outer-gutter/2;
+  @include media($large-screen) {
+    padding-left: $outer-gutter;
+  }
+  @include display(flex);
+  @include flex-wrap(wrap);
+  @include align-items(center);
+  padding-top: rem(5.5);
+  padding-bottom: rem(5.5);
+  font-size: rem(16);
+  font-weight: $font-semibold;
+  line-height: rem(22);
+  .covid-19-vaccine-link__picture__container {
+    height: rem(50);
+    width: rem(50);
+    padding: rem(5);
+    display: flex;
+    margin-right: rem(5);
+    background-color: white;
+    border-radius: 50%;
+    img { // ie hack
+      flex-shrink: 0;
+    }
+  }
+  .covid-19-vaccine-link__button {
+    @include button-unstyled;
+    font-weight: $font-semibold;
+    cursor: pointer;
+    text-decoration: underline;
+    &:hover {
+        color: $color-primary;
+    }
+  }
+}
