+++ src/_scss/pages/homepage/homePage.scss
@@ -7,7 +7,7 @@
     .homepage-content {
         background-color: $color-white;
     }
-
+    @import "./covid-vaccine-link";
     @import "./hero/hero";
     @import "./features/features";
     @import "./download/download";
